Move default Expo components to expo_components dir
