+++ components/__tests__/ThemedText-test.tsx
@@ -1,10 +1,12 @@
-import * as React from 'react';
-import renderer from 'react-test-renderer';
+import * as React from "react";
+import renderer from "react-test-renderer";
 
-import { ThemedText } from '../ThemedText';
+import { ThemedText } from "@/components/expo_components/ThemedText";
 
 it(`renders correctly`, () => {
-  const tree = renderer.create(<ThemedText>Snapshot test!</ThemedText>).toJSON();
+  const tree = renderer
+    .create(<ThemedText>Snapshot test!</ThemedText>)
+    .toJSON();
 
   expect(tree).toMatchSnapshot();
 });
